Use absolute path to call psalm (#3)
Since three weeks, calling `psalm` in the `entrypoint.sh` script doesn't
work anymore, the binary is not found, as though `PATH` is expanded to
`/composer/vendor/bin` in the `Dockerfile`. I hove no idea why that is,
but it should be better to call `/composer/vendor/bin/psalm` by it's
absolute path anyways.